+++ CHANGE_LOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added "postbin" command which allows piping payloads from Postb.in to a URL in the same manner as GitHub webhooks
 
 ## [0.1.1]
 ### Changed